intro/pyproject.toml
[tool.poetry]
name = "intro"
version = "0.1.0"
description = ""
authors = ["esilean <le.bevilaqua@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^1.10.8"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
requests = "^2.31.0"
aiohttp = "^3.8.4"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
